requirements.txt
google-cloud-bigquery
docker
ipinfo